Handle opening of non-existant files better in libcloud storage
Currently, the LibCloudStorage._read method raises the following exception when the file doesn't exist:

    'NoneType' object has no attribute 'size'

This is because the "obj = self._get_object(name)" line is returning None, but no check is made on "obj" before using it.

This commit changes this behaviour so it now checks if _get_object returned None and instead raises a LibCloudFileNotFounderror (a subclass of IOError).

The file is now downloaded when it is opened. This is so the exception is raised where the developer would expect it