Add and use _normalize_name() like in s3boto
We need this even though we don't have "location" support since it undoes
some of the weirdness (e.g. adding './' at the beginning of some paths)
done by clean_name().

It would be possible to do this more simply but doing it this way makes
things consistent with s3boto.