Update to Amazon S3 Documentation
Included note indicating that AWS_S3_CUSTOM_DOMAIN must *not* end in a slash. It is required that `STATIC_URL` must end in a trailing slash, and if users are using CloudFront, their `STATIC_URL` and `AWS_S3_CUSTOM_DOMAIN` values will be the same. It is very easy for users to miss the trailing `/` and instead set `AWS_S3_CUSTOM_DOMAIN = STATIC_URL`.