Avoid incorrect behaviour of S3BotoStorage.exists
Calling `S3BotoStorage._save` writes to `_entries` if `preload_metadata`
is True. Future access to `_entries` is limited as the property `entries` only
loads metadata for the bucket if `_entries` is empty, which due to `_save`
it may not be. This in turn causes `exists` to incorrectly return False for
existent files.