Avoid silently changing settings; raise ImproperlyConfigured instead
If a setting is considered invalid, raise an exception instead of
silently changing it to some other value. Avoids confusion and forces
the library user to be explicit.

For example, setting S3 location to a value with a leading slash
previously stripped it. Instead, now instruct the user to use a valid
value without the leading slash.