Remove confusing sentence about legacy storage s3.S3Storage
While reviewing the docs, I found the reference to "extreme legacy
backend" confusing and ambiguous. It provides no details and doesn't
name what was removed. Diving into the git logs, I discovered this was
referencing a previous S3 backend, "storages.backends.s3.S3Storage". As
this was removed in 2014 and was part of the quite old 1.2 release, I
think enough time has passed such that this reference can be completely
removed.