Calculate MD5 for each S3 multipart upload incrementally
If ContentMD5 is not passed, botocore process the entire file to work
out the MD5 for the upload to S3:

https://github.com/boto/botocore/blob/1ae092ed95b39c3336ba48c5cbd71b7c37924d05/botocore/handlers.py#L200

It's likely better in many situations to calculate the MD5 for each set
of bytes as they come in: the data is already in memory/caches.

For event-loop based applications, where file io is often still
blocking, this is especially valuable: botocore would block the event
loop when it tried to calculate MD5.