Allow S3Transfer config to be customised
A typical use for this would be to control number of threads, or even
disabling threads. For example, some transfers are unlikely to be faster
than the transfer to/from S3, and so triggering threads is unlikely to
be helpful.

The setting AWS_S3_TRANSFER_CONFIG is used twice so that S3Boto3Storage
and S3Boto3StorageFile, which can be used directly in client code, can
both be customised.