Use skip_install=true for flake8 tox target
Avoids installing the package (and any potential dependencies) to the
virtualenv before running flake8 commands. The package is not required
to be installed to do simple static code analysis. Results in a slightly
faster run, as fetching and installing dependencies is skipped.

For additional information on the configuration option, see:

https://tox.readthedocs.io/en/latest/config.html#confval-skip_install=BOOL

> Do not install the current package. This can be used when you need the
> virtualenv management but do not want to install the current package
> into that environment.