Convert string to bytes if needed
When django-storages is used from an app which is configured by env variables
fed into dynaconf (or probably other situations), it's not easy to directly
coerce a string representation of the signing key into a bytes-like object; it
comes in as a string.  This detects that situation and encodes the string.