S3: Avoid a ListParts request in multipart upload
S3 returns the ETag for each part of the upload while we're PUTting
them. However, until now we were ignoring those responses, and instead
doing a separate ListParts (GET) request for the upload object, just
before POSTing the list of parts to the completion endpoint to finalise
the upload.

That extra GET request is unnecessary so this commit removes it.