Fix install instructions
The install instruction needs to be updated. The azure package is deprecated and broken down into separate packages. In addition, the new django-storages doesn't work with the old azure package due to this issue (https://stackoverflow.com/questions/50581138/cannot-import-name-blockblobservice). 

I propose this new update to the installation process. I have tested this process and it works perfectly. Feel free to reach out to me if you have any questions.