[sftp] Improve write & memory performance when saving files (#1194)
* Move seekability test to a separate helper function

Having a third copy of this code would feel silly.

* Stop calling content.open() from SFTPStorage._save

The open method isn't available in all file-like objects one might want to
save using the storage framework, so instead of relying on it calling
seek(0) for us do it ourselves (conditionally).

* Require at least version 1.10.0 of paramiko

This version introduced the SFTPClient.putfo() method. It was released in
early 2013, so this shouldn't be a great hardship.

* Use SFTPClient.putfo() in SFTPStorage._save()

Instead of using SFTPClient.open() to get a file-like object, reading the
entire content into memory and then calling .write() on that.

The measured performance difference is stunning, and the uploaded file data
no longer needs to fit into memory.