Introduce S3SeekableFile as a S3 file wrapper
Previously, whenever an S3 file was opened, it was directly downloaded
into a `SpooledTemporaryFile`, independently if it was ever read or not.
For files with a bigger size it reduced performance and increased
traffic and expenses in cloud environments.

This commit wraps the access to the S3 file into a S3SeekableFile
wrapper which reads requested bytes on demand, without loading the file
completely into memory or local storage.